Check if ./.hg/ and ./.git/ are directories
Summary:
We're discovering the `.arcconfig` file by looking into a few
directories, one of which is `./.git/` or `./.hg/`. If in current
directory a `./.git` or `./.hg` file has been created, an `IOError` `ENOTDIR`
is raised. The `./.git` file is part of any git worktree, and the user
might accidentally run moz-phab from such directory.

This patch is checking the way we find out if a path is a "dot path" one.
The `__init__()` method of `Mercurial` and `Git` classes is now not
only looking if the path exists, but also if it is a directory.

Test Plan:
Create a directory with a .git or .hg file in it.
Try to submit from this directory before and after applying this patch.

Reviewers: mars, glob

Reviewed By: glob

Subscribers: tromey

Bug #: 1494968

Differential Revision: https://phabricator.services.mozilla.com/D7596