Check if arc is configured before proceeding any real action.
Summary:
We're calling arc after the checkout. If arc fails the repository is in
a different state than before calling the `moz-phab`.
We should check if `arc` is running correctly before proceeding with the
submit.

This patch is sending a ping to Phabricator's API.
We cache the call by creating a file in the repository dot path.

Test Plan:
Run moz-phab with unconfigured arcanist and observe it failing with a message.
Install the certificate.
Run moz-phab (No need to perform an actual submit) and check if file `./.git/.moz-phab_arc-configured` exists.

Reviewers: glob

Reviewed By: glob

Bug #: 1496166

Differential Revision: https://phabricator.services.mozilla.com/D7987