Bug 1496029 - Handle arcanist as a system package r=glob
When arcanist is installed as a system package, attempting to update
`moz-phab` (either through `moz-phab self-update` or through the periodic
update) will fail due to `arc upgrade` failing. This failure is caused
by `arc` not being located in a checkout of the arcanist repo. When this
occurs, we now prompt the user if they would like to disable calls to
`arc upgrade` in the future. Even if they decline, the update process
will continue and `moz-phab` will update itself.

Differential Revision: https://phabricator.services.mozilla.com/D13308