Bug 1510279 - (HG) Do not rebase if parent is not changed. r=glob,dylan
Summary:
If no change has been done in the parent (part of the stack
resubmitted) there is no need for rebase commit on the new parent.
Mercurial would raise an exception and stop moz-phab execution.
This patch reads the real parent and omits the rebase step if it's
the same as the parent in the commit stack.

Reviewers: glob, dylan

Reviewed By: dylan

Bug #: 1510279

Differential Revision: https://phabricator.services.mozilla.com/D14544