Improve handling of temporary files for cross-platform compatibility
Summary:
This patch adds a TemporaryFileName() context manager that takes
a string and writes that string to a temporary file.
The temporary file will exist for the duration of the context
and can be passed to other programs for reading.

This change is needed because (without C-level API calls that are not exposed to
python) it is not possible for multiple processes to both have an open handle on
a file at the same time.

This change also removes the open() and seek() behavior used for calling arc,
instead of using a temporary file and a call to open(). This was done to make
the code simpler, allowing for the simpler API of TemporaryFileName().

Test Plan: I'll ask zalun how to run through a full text with both VCSes on Windows.

Reviewers: zalun

Reviewed By: zalun

Bug #: 1514906

Differential Revision: https://phabricator.services.mozilla.com/D14872