Bug 1520155 - Rebase feature branches in Mercurial with evolve extension. r=glob
Summary:

`moz-phab` is rebasing only the commits within the commit stack. If any
of the commit is a DAG branch point and is amended during the submit
process, the feature branch will not be rebased and will be left
attached to the obsolete commit.

With this patch `moz-phab` is storing the commits children when
building the commit stack and is using this info to rebase feature
branches in the `finalize` stage.

DAG branch points ca't be in stack if no evolve extension is installed.

Reviewers: glob

Reviewed By: glob

Bug #: 1520155

Differential Revision: https://phabricator.services.mozilla.com/D16558