Bug 1495777 - Allow to use full VCS configuration. r=glob
Summary:

Currently all calls to VCS are done with a limited configuration.
This patch is changing that and runs the Git and Mercurial using full
config. `--safe-mode` can be used as an option. If user wants to always
run the VCS commands in a safe mode `vcs.safe_mode` config item is added.

Reviewers: glob

Reviewed By: glob

Bug #: 1495777

Differential Revision: https://phabricator.services.mozilla.com/D17944