Bug 1527621 - Add ability to patch without creating a branch in Git r=glob
Summary:
Mercurial has a `--no-bookmark` option which allows to not create a
bookmark for a new DAG branch point. In Git we had been always create a
branch.

This patch is adding a new option `--no-branch` which allows to create
commits without creating a branch. This may result in switching the
repository to the 'detached HEAD' state. User has to use the `--yes`
option or agree to the prompt.

Reviewers: glob

Reviewed By: glob

Bug #: 1527621

Differential Revision: https://phabricator.services.mozilla.com/D19773