Bug 1591015 - CRLF allowed in git and hg r=nika,glob!
Python was stripping the `\r` from the diff response.

Warning:
Please take a look at detecting the  `\\ No new line at end of file\n`
string. There is a change as we now have the `\n` character at the end
of this line. We might want to change it.

Warning:
There is no test for update yet. Creation was working fine even before
the fix.

Note:
recognizing lack of the `\n` at the end of the file was broken.

Differential Revision: https://phabricator.services.mozilla.com/D50958