Bug 1600696 - use non-arc implementation by default, adds --arc r?zalun
I've tried to do this with as few modifications as possible. I _could_ replace all instances of no_arc with arc within the codebase, but since we're going to be deleting the arc code soon enough, I figured that it's not worth the potential breakage.

Note that the tests are remaining the same (if they didn't use arc before, they still don't, but if they did use arc before, they still use arc now).

Differential Revision: https://phabricator.services.mozilla.com/D55757