No bug - allow "--trace" with any moz-phab command
This could be solved by adding "--trace" to the top-level parser, but then you'd have to run moz-phab like:
"moz-phab --trace patch ..."
If you appended "--trace" to the end, the top-level parser wouldn't understand it.

So, I've added "--trace" to each subcommand, which is a bit of a bummer, but an improvement overall.

Additionally, I resolved an issue with `logger.error(...)` when DEBUG was True (a mix of debug and non-debug args were being provided)

Depends on D57916

Differential Revision: https://phabricator.services.mozilla.com/D57921