Bug 1675648 - improve handling of line separators r=sheehan,dkl
This change improves line separator support on different operating
systems by:

- Determining linesep character based on current OS (e.g. `\n` vs `\r\n`)
- Splitting lines that are generated by `hg cat` only on linesep
    - This has the side-effect of preventing stray characters such as
    carriage returns from triggering a line split
- Combining shared functionality into a new `generate_lines` helper method

The behaviour prior to this change was to use the `str.splitlines`
method, which splits lines on a number of characters (see
https://docs.python.org/3/library/stdtypes.html#str.splitlines). This
leads to corrupt hunks when characters other than those used as line
separators are used while splitting.

The prior behaviour would also have caused some inconsistencies when
generating hunks on Windows, particularly when a newline character was
missing from the end of a file (as in that case, a carriage return
character would have been missing on the last two lines in the hunk).

The behaviour of `str.splitlines` is very particular, and is quite distinct
from the behaviour of `str.split`. `str.split` also behaves differently
when a delimiter is provided, vs. when the default `None` is used. In
the former case, an empty string will be included in the result of the
split, but in the latter case, an empty list will be yielded. See
https://docs.python.org/3/library/stdtypes.html#str.split for more info.

The implementation of `generate_lines` takes into account all the above,
as well as the behaviour that was achieved by the `keepends` parameter,
which does not exist in `str.split`.

Differential Revision: https://phabricator.services.mozilla.com/D101215