uplifts: add a `moz-phab` uplift command (Bug 1737838)
Add a `moz-phab uplift` command to send commits to Phabricator
to be reviewed as uplift requests. The interface is similar to
`moz-phab submit`, in that a revset/commit range is specified
for submission, but the user must also pass a `--train` argument
with a Phabricator callsign indicating the target uplift train
to submit to. The target repos are any repository on Phabricator
with an `uplift` project tag associated with the repository. If the
repository has an appropriate named tag/bookmark/branch associated
with the passed train, a new set of revisions will be created via
rebase and the new commits will be submitted for uplift. When amending
the commits, the `Differential Revision` of the original revision
is moved to `Original Revision` and the uplift revision is set as
the `Differential Revision`.

To keep uplifting behaviour consistent with regular submissions the
`uplift` command re-uses almost all of the logic of `submit`, with
a few amendments that check the value of `args.command` to determine
if the uplift specific code paths should be taken. The `submit`
CLI argument parser is turned into a function so it can be re-used
by the `submit` parser.

A conduit function call is added to return repositories with an
associated tag, in our case the `uplift` tag. We add a few VCS specific
functions to associate a callsign with a named branch/symbol and to
perform the automatic uplifting functionality.

TODOS
    - no telemetry
    - if rebase fails we continue and submit anyways, probably wrong
    - needs unit tests