circleci: add workflows to publish releases on pushed tags (Bug 1761982) r=zeid
Adds a new job to publish a release of `moz-phab` to PyPI when
a tag is pushed to the repo. Adds a simple set of instructions
on how to create a release to `README.md`. Add a simple check
that if `CIRCLE_TAG` is present in the environment, it matches
the value of `VERSION` in `setup.py` to avoid headaches if tags
are pushed without updating the version in lockstep.

Differential Revision: https://phabricator.services.mozilla.com/D160551