uplifts: avoid rebasing descendant commits and clarify repo amend logic (Bug 1757076) r=zeid
Adds a check to see if the repo revset is a descendant of the
unified head on which the rebase would apply. In the Mercurial
case we query the log to check if the submit revset is descendant
from the unified head. For Git we use the `merge-base` plumbing
command and the `--is-ancestor` argument, as this is the inverse
of finding a descendant.

Since we are adding logic that changes the repo amending conditions
we take this time to clarify that logic, moving everything into a
standalone function that returns a boolean indicating if the later
submit actions should make local amendments to the revset after
submission.

Differential Revision: https://phabricator.services.mozilla.com/D140207