conduit: preserve revision status during re-org (Bug 1731811) r=zeid
When `moz-phab` submits a revision as `wip` it makes two subsequent
calls to `differential.revision.edit` Conduit endpoint. The first call
create/updates the revision information, and the second calls sets
the `changes-planned` status. This is a workaround for Phabricator
throwing an error when attempting to set a status to the same value
on a revision, combined with Phabricator setting the status to
`needs-changes` on update.

Extend this `set_wip_later` call concept to be a set of `post`
transaction calls and make `--wip` use it. Then add a `preserve_status`
flag which adds a `post` transaction to set the revision status back
to it's original value after the first update. To do this we create a
mapping of known API status values to transaction type keys, since the
values are not the same string. Make `moz-phab re-org`'s calls to
`edit_revision` preserve the status, solving the issue of `moz-phab
reorg` erroneously setting status to `needs-review`.

Differential Revision: https://phabricator.services.mozilla.com/D142330