user: account for an empty `primaryEmail` field in Conduit response (Bug 1651898) r?zeid
It is possible for a user to have an empty `primaryEmail` field
in their Conduit response. Account for this potential state,
logging a warning that advises them to add a primary email in
their Phabricator settings. If they don't have a primary email
set, assume they are not an employee and return.

Differential Revision: https://phabricator.services.mozilla.com/D142487