ci: add workflows and support for Python 3.10 (Bug 1762438) r=zeid
Adds a CircleCI job for Python 3.10 testing, a `docker-compose`
command to generate requirements for Python 3.10 and the result
of running the command. Amend `README.md` to include the new
command.

The current versions of `black` and `pytest` fail to install on
Python 3.10, so upgrade them in all requirements files. For `black`
this also requires a slight syntax change to
`tests/test_integration_git.py`, so add those syntax changes.

Differential Revision: https://phabricator.services.mozilla.com/D142586