telemetry: make new and existing UUIDs conform to expected Glean schema (Bug 1788719) r=grammar
Add code to detect and rewrite existing 32-character non-hyphenated
UUIDs to 36-character hyphenated UUIDs. Also generate new UUIDs
with the correct format for new clients.

Some tests were rewritten to conform to the new UUID format. One
test was difficult to mock so it has been rewritten to check the
expected properties of the returned values instead of expecting
an exact match.

Differential Revision: https://phabricator.services.mozilla.com/D156852