self-update: rework update checking to make self-update display some information (Bug 1798751) r=zeid
Re-work `check_for_updates` to allow re-using the logging
logic in `moz-phab self-update`. Control running the self
update from the update checker via a function argument instead
of a config knob so we can always force the update when
running the update command explicitly. Move the restarting
of `moz-phab` to its own function outside of the updating
logic so it is only run when doing an update check outside
of `moz-phab self-update`. Update mocks of `check_for_updates`
to have a `return_value=None`, which also removes requiring
them as function arguments in tests.

While here, reverse the `if` logic in `check_for_updates`
to exit if the update conditions aren't met, allowing the
entire function to be de-dented a level. Clarify the `if`
statement with clearer variables and add a `force_check`
kwarg to override our environment based checks when running
`moz-phab self-update`.

Differential Revision: https://phabricator.services.mozilla.com/D161170