updater: use `sys.argv[0]` to resolve script dir (Bug 1774074) r=zeid
Using `sys.path[0]` to resolve the script directory works on
Linux/MacOS, however on Windows the full path to the starting
script is passed as `sys.path[0] instead of the directory to
the script.

Change how we define `script_dir` to using `sys.argv[0]` which
is the path to the script used to start Python. The parent
directory to this value is the script directory.

Differential Revision: https://phabricator.services.mozilla.com/D164514