args: add diff ID flag to patch command (Bug 1534707) r=sheehan
Right now users cannot specify a diff ID within a revision's history
when patching. Being able to specify a diff allows for investigating
older diffs and provides and easy way to rollback.

This change lets users use `--diff-id` to supply a number as long as
it is a Diff ID that belongs to a revision in the stack, it will
apply changes from the diff to the local repo being patched.

This will retrieve the diff via ID and replace the existing diff info
for that revision with the requested diff's info.

Differential Revision: https://phabricator.services.mozilla.com/D165314