uplift: use repo shortnames instead of callsigns for `--train` (Bug 1809214) r=zeid
The current implementation of the `train` parameter in `moz-phab` uses
the repo callsign to specify the uplift target. The `shortname` is
what the web UI uses and is the nicer name value we usually refer
to repos with.

Rename `get_repository` to `get_repository_by_callsign` to clarify
the input parameter and add a `get_repository_by_shortname` function.
Add a helper function that queries Phabricator for a single repository.
Update `list_trains` to print the shortnames of the uplift repos instead
of the callsign. Then update `map_train_arg_to_repo` to use the
shortname to find the repo information on Phabricator and have the
callsign set from the returned information instead of from
`args.train` directly.

Differential Revision: https://phabricator.services.mozilla.com/D166458