conduit: remove preserve status functionality (Bug 1773145) r=zeid
Remove the busted `preserve_status` functionality. The idea behind this
code was to have a post-operation transaction that reset the status
of a revision to the previous state before editing the revision.
In practice this code hit many edge cases that made it unusable
(you cannot accept your own revision, for example). We remove
the `preserve_status` flag as well as the revision state to
transaction mapping. We keep the `post_trans` functionality instead
of reverting to a flag specifically for `wip` as the model
of a "post edit transaction" more closely resembles our intended
use case.

Differential Revision: https://phabricator.services.mozilla.com/D166931