reorg: avoid abandoning revisions that are already abandoned (Bug 1626028) r=grammar
`moz-phab reorg` fails to properly re-org a revision stack
when there is a revision in the abandoned state that we
are attempting to remove from the stack.

Add a second revision fetch to retrieve remote revision
information for patches in the stack which are missing
on our local stack. Then pass the set of abandoned
revisions through to `stack_transactions` to avoid
adding an "abandon" transaction to the set of changes
to make for that revision.

Differential Revision: https://phabricator.services.mozilla.com/D167182