bmo: only default to mozilla bugzilla if phab url is mozilla (Bug 1822104) r=sheehan
For non-mozilla projects that doesn't specify a bmo URL in its
.arcconfig moz-phab defaulted to using https://bugzilla.mozilla.org for
bmo url.

This was problematic as it caused the phabricator API token for the
non-mozilla project to be sent to mozilla's bugzilla.

This commit changes so that the default is calculated when setting up
Repo object rather than when doing request. This allows additional logic
so that it only is set to bugzilla.mozilla.org if the phabricator url
looks like a mozilla owned project.

Differential Revision: https://phabricator.services.mozilla.com/D172540