patch: make branch name configurable (Bug 1826125) r=zeid
Adds a `patch.branch_name_template` config knob to control
the name used as the branch name for a new bookmark/branch.
This knob takes a single format template, `rev_id`, which
is the rev id of the revision being patched. The current
default value is `phab-D{rev_id}` and that is set as the
default value in the config. Move branch name resolution
into a separate `resolve_branch_name` function and add a
test for it.

Differential Revision: https://phabricator.services.mozilla.com/D174777