styling: add `noqa: B019` to lru_cache calls (Bug 1870291) r=zeid
Add a `noqa` for rule B019 to uses of `lru_cache`. Using the
lru cache in this way is a problem as the `self` value is used
as a cache key, which means a reference to the instance is kept
around by the cache, which can cause a memory leak since the
instance can't be garbage collected.

In our case, there is only ever one instance of a `Repository`
for the lifetime of the program, so we aren't worried about
an issue with the garbage collector.

Differential Revision: https://phabricator.services.mozilla.com/D196792