git: rename `commit` loop variable to `stack_commit` to avoid double alias (Bug 1870284)
When creating the `Commit` object I renamed the `c` variable name
to `commit` to keep the variable names consistent and readable.
In the `amend_commit` function we already have a `commit` variable
defined in the function signature, effectively overwriting the
variable name and causing some issues with rebasing commits during
rewrite. Change the loop variable to `stack_commit` to reflect that
this variable is the current `Commit` object we are looping through
in the stack.