updater: reverse logic for self-update and move into a separate function (Bug 1833395) r=zeid
The logic for when a self-update should occur is incorrect, and causes
the check for self-updating to return early instead of correctly
checking for a new moz-phab version. The existing `self_update_disabled`
variable should have been checked for the `False` case instead of
the `True` case in the existing code structure.

Move the logic into a separate function that returns a boolean indicating
if a self-update should be performed and logs the reason for why a
self-update is skipped. Add a test for this new function to verify the
self-update logic is working as expected.

Differential Revision: https://phabricator.services.mozilla.com/D197189