hg: move `--config` so `rebase` is the first command arg (Bug 1879129) r=zeid
When `hg_out` fails to run a command, the message that is used for
the exception is the a "command" failed, where the command is the
0th element of the command arguments. Since we pass the `--config`
option before the `rebase` command in `uplift_commits`, this causes
the error message to indicate a failure of the `--config` command.

Move the `--config` flag to the end of the command, so the actual
`rebase` command ends up in the exception error message.

Differential Revision: https://phabricator.services.mozilla.com/D201102