spinner: add an environment variable to enable/disable the spinner (Bug 1882575) r=zeid
`moz-phab`'s spinner implementation has a bug which can cause hangs.
While we track down the bug, it would be convenient to add an escape
hatch to disable the spinner, similar to other similar variables. Add
a `DISABLE_SPINNER` environment variable that can be used to hide the
spinner when set in the environment.

Differential Revision: https://phabricator.services.mozilla.com/D203037