submit: fix `test_submit_update_revision_not_found` (Bug 1887834) r=sheehan
In some scenarios that are exposed by the next patch, `get_diffs`
was called with an empty list in `show_commit_stack`, which raises an
error. This can happen when a commit is submitted with a revision ID
that doesn't exist.

Fix assert in `test_show_commit_stack` that depended on previous state.

Differential Revision: https://phabricator.services.mozilla.com/D224093