commits: move commit message building into `Commit.build_commit_message` (Bug 1924899) r=shtrom
In `moz-phab submit` we have some code which builds template variables
and passes them to `arc_message` to build a commit message for the
given commit. Since each of the fields used to build the commit message
are in the `Commit` object, we move this functionality into a
`build_commit_message` function on the `Commit` object.

`build_commit_message` requires `revision_title_from_commit`, but
since that function is also built entirely from `Commit`, we also
move it into the `Commit` object.

Since `revision_title` requires the `WIP_RE` regex object we move
it into `commits.py`, and since the only other consumer of that
regex is `wip_in_commit_title`, which is also entirely built from
attributes of a `Commit`, we also move `wip_in_commit_title` into
`Commit`. `wip_in_commit_title` is previously tested with raw strings,
so we update the test to build a `Commit` first instead.

Differential Revision: https://phabricator.services.mozilla.com/D225735