commits: set revision PHID when validating after loading from Phab (Bug 1883957) r=shtrom
In `show_commit_stack` with `validate=True`, we take the list of
commits and load the associated revision data from Phabricator.
`show_commit_stack` then sets various flags, and iterates through
the commits to display warnings to the user. One of the flags
that is set is `commit.submit`, which is a boolean that controls
if the commit should be re-submitted or skipped due to being a
no-op update. When validating commits from loaded Phabricator API
data, we currently do not set `commit.rev_phid`, on the `commit`,
which means it is only ever correctly set after creating a new
revision in `moz-phab submit`. When a new revision is added to the
top of a stack, and the parent commit is detected as a no-op submit,
the `rev_phid` is never set on that revision and thus the new commit
will not have a `parent_phid` passed to ensure it is correctly
associated with the stack.

Populate `commit.rev_phid` with the PHID returned from Phabricator
during validation in `show_commit_stack`.

Differential Revision: https://phabricator.services.mozilla.com/D226504