testing: add a test that enforces bug numbers in commit messages (Bug 1928130) r=glob
Add a test that checks all commits since `origin/main` that touch code
(ie not documentation or tests) have bug numbers in their commit
messages. This will enforce the bug number requirement in CI, and will display
both the commit SHA and the commit message title in the failing test.

Differential Revision: https://phabricator.services.mozilla.com/D227377