git: improve base commit detection and messaging (Bug 1966122) r=shtrom
Switch `_get_first_unpublished_node` to use `mach try`'s base
detection method of looking for the first node in the history
up to the current HEAD which is not present on a remote. This
avoids `moz-phab` detecting `origin/main` as the upstream to
compare against when the user has used a different branch (ie
`autoland`) to base their patch against, resulting in revisions
which already exist in the repo being gathered for submission.
After this change, `moz-phab` will automatically detect commits
from any remote branch as published.

We add some logic for detecting the appropriate remote to use,
defaulting to the `git_remote` config setting first, then looking
for `origin` if it exists, and using all found remotes if it
is absent. In each case we log a message about how the remote
was detected and direct users to the `git_remote` setting
to override the default behaviour.

Since `_cherry` is no longer used after this change, it is removed.
We also move the check for over 100 detected commits to the `submit`
command definition directly.

Differential Revision: https://phabricator.services.mozilla.com/D254557