git: use `get_base_remotes` when building repo branch head (Bug 1984467) r=shtrom,mhodson
Currently we hard-code `origin` as the remote name when finding
a remote branch to use as the uplift target to rebase patches
onto. Instead use the result of `get_base_remotes` to build
the remote branches. The messaging in `get_base_remotes` will
inform the user of multiple remotes being available and suggest
the `git_remote` config option to enforce a specific remote in
the event we select the wrong remote branch. Update wording in
one log message to remove references to finding a base commit,
as that was the single usage of this method previously. Also
update tests to show the expected behaviour change after this
patch.

Differential Revision: https://phabricator.services.mozilla.com/D262193