Bug 1988631 - fix(jujutsu): avoid revset syntax errors w/ `{start}::{end}` r=sheehan
This improves diagnostics with revsets that may otherwise be
syntactically invalid. For instance, a user might instinctively specify
`<start>::<end>` instead of two separate `start` and `end` CLI
arguments, which would produce this error diagnostic:

```
CommandError: command 'jj' failed to complete successfully
Run moz-phab again with '--trace' to show debugging output
```

With this change, the diagnostic changes to:

```
Multiple parents found for change <snip>, unable to continue
```

…which does a much better job of hinting that the argument is expected
to unambiguously refer to a single change, rather than refer to a range.

Differential Revision: https://phabricator.services.mozilla.com/D264885