uplift: display Lando URL for the new revision on successful `moz-phab uplift` (Bug 1988251) r=shtrom
After bug 1987196, the uplift request assessment in Phabricator
has been removed and Lando must be used to respond to the form.
Display a link to the Lando URL for the revision on successful
`moz-phab uplift`.

Update the `submit` command handler to return the list of commits
it generated, and use the tip commit in the stack to build the
Lando URL. Add a mapping of known Phabricator domains to their
associated Lando domains and provide a helper method on the
repository object to get the Lando URL.

Differential Revision: https://phabricator.services.mozilla.com/D264642