uplift: send developers to new uplift linking URL (Bug 2006722) r=shtrom
Switch the displayed Lando URL at the end of a successful
`moz-phab uplift` from the tip revision Lando view to the new
assessment batch linking URL.

Add a new `--assessment-id` argument to `moz-phab uplift` that can
be used to directly link revisions to the expected assessment form
after submission. The argument will be added to the resolution steps
hint in the Lando UI after it is released.

Differential Revision: https://phabricator.services.mozilla.com/D276909